Fixxing pro -m

+++ src/test/test.sh
@@ -20,11 +20,17 @@ if [ `$wc $a.out -l | cut -d  ' ' -f 1` -eq $size ]; then
     then
         commands[$a]="PASS"
     fi
-    rm -f $a.out
+    if [[ "$CANNON" == true ]]
+    then 
+	cp $a.out ./output/${msg}.out
+	rm -f $a.out
+    else
+	rm -f $a.out
+    fi
 else
-	echo `$wc $a.out -l `
-	echo "expected $size"
-	echo "[$a] $msg FAIL"
+    echo `$wc $a.out -l `
+    echo "expected $size"
+    echo "[$a] $msg FAIL"
     commands[$a]="FAIL"
     if [[ "$HALT" == true ]]
     then
@@ -38,7 +44,43 @@ else
 fi
 }
 
+testb() 
+{
+if [ $? -ne 0 ]; then
+	echo "[$a] produced unexpected behavior."
+	exit 1
+fi
+
+if ! [ "$(diff -q $a.out ./output/$msg.out)" ]; then
+	echo " PASS"
+    if [[ ${commands[$a]} == "UNTESTED" ]]
+    then
+        commands[$a]="PASS"
+    fi
+    rm -f $a.out
+else
+    echo "$(diff $a.out ./output/$msg.out)" 
+    echo "[$a] $msg FAIL"
+    commands[$a]="FAIL"
+    if [[ "$HALT" == true ]]
+    then
+        echo "If you would like to continue testing after this error call './test.sh HALT=false"
+        cat $a.out
+        exit 1
+    else
+        echo "If you would like to halt at this error call './test.sh HALT=true"
+        rm -f $a.out
+    fi
+fi
+}
+
+testc()
+{
+testa
+}
+
 HALT="${HALT:-true}"
+CANNON="${CANNON:-false}"
 
 header="spitze-header.txt"
 mpileup="spitze-mpileup.txt"
@@ -51,7 +93,7 @@ unicode="name-file-unicode.txt"
 
 formats=($mpileup)
 
-pop=8			#number of individuals in population
+pop=8		#number of individuals in population
 idx_header=3	#Size of pro header
 pro_header=3	#Size of pro header
 vcf_header=3	#Size of vcf header
@@ -81,45 +123,54 @@ do
 done
 
 a="proview"
-msg="proview"
+msg="proview_1"
 rm -f $a.out
 size=$(($pro_size+$pro_header+$idx_size+$idx_header+$smp_size))
 echo -n "cat $mpileup | $mapgd $a -H $header > $a.out							"
 $timeout 5s bash -c "cat $mpileup | $mapgd $a -H $header > $a.out"
-testa
+testc
 
 a="proview"
-msg="proview"
+msg="proview_1s"
+rm -f $a.out
+size=$(($pro_size+$pro_header+$idx_size+$idx_header+$smp_size-2))
+echo -n "cat $mpileup | $mapgd $a -s -H $header > $a.out							"
+$timeout 5s bash -c "cat $mpileup | $mapgd $a -s -H $header > $a.out"
+testc
+
+a="proview"
+msg="proview_2"
 rm -f $a.out
 size=$(($pro_size+$pro_header+$idx_size+$idx_header+$smp_size))
 echo -n "$mapgd $a -n $name -H $header > $a.out								"
 $timeout 5s bash -c "$mapgd $a -n $name -H $header > $a.out"
-testa
+testc
 
 a="proview"
-msg="proview"
+msg="proview_3"
 rm -f $a.out
 size=$(($pro_size+$pro_header))
 echo -n "$mapgd $a -n $name -H $header -o $a; ($a.pro)							"
 $timeout 5s bash -c "$mapgd $a -n $name -H $header -o $a"
 mv $a.pro $a.out
-testa
+testc
 
+msg="proview_4"
 size=$(($idx_size+$idx_header))
 echo -n "$mapgd $a -n $name -H $header -o $a; ($a.idx)							"
 mv $a.idx $a.out
-testa
+testc
 
 a="allele"
-msg="allele"
+msg="allele_1"
 rm -f $a.out
 size=$(($pro_size+$pro_header+$idx_size+$idx_header+$gof_header+$gof_size))
 echo -n "$mapgd proview -n $name -H $header | $mapgd $a -c 1 > $a.out				"
 $timeout 5s bash -c "$mapgd proview -n $name -H $header | $mapgd $a -c 1 > $a.out"
-testa
+testc
 
 a="allele"
-msg="allele"
+msg="allele_2"
 rm -f $a.out
 mkfifo out.pro
 size=$(($pro_size+$pro_header))
@@ -127,8 +178,7 @@ echo -n "$mapgd proview -n $name -H $header | $mapgd $a -c 1 -p out > /dev/null	
 $timeout 5s bash -c "$mapgd proview -n $name -H $header | $mapgd $a -c 1 -p out > /dev/null &"
 cat out.pro > $a.out
 rm -rf out.pro
-testa
-
+testc
 
 a="pool"
 msg="pool"
@@ -136,7 +186,7 @@ rm -f $a.out
 size=$(($pro_size+$pro_header+$idx_size+$idx_header))
 echo -n "$mapgd proview -n $name -H $header | $mapgd $a > $a.out						"
 $timeout 5s bash -c "$mapgd proview -n $name -H $header | $mapgd $a > $a.out"
-testa
+testc
 
 a="filterpool"
 msg="filterpool"
@@ -144,43 +194,43 @@ rm -f $a.out
 size=$(($good_pool+$pro_header+$idx_size+$idx_header))
 echo -n "$mapgd proview -n $name -H $header | $mapgd pool | $mapgd $a > $a.out	"
 $timeout 5s bash -c "$mapgd proview -n $name -H $header | $mapgd pool | $mapgd $a > $a.out"
-testa
+testc
 
 a="filter"
-msg="filter"
+msg="filter_1"
 rm -f $a.out
 size=$(($good_size+$pro_header+$idx_size+$idx_header))
 echo -n "$mapgd proview -n $name -H $header | $mapgd allele -c 1 | $mapgd $a > $a.out	"
 $timeout 5s bash -c "$mapgd proview -n $name -H $header | $mapgd allele -c 1 | $mapgd $a > $a.out			"
-testa
+testc
 
 a="filter"
-msg="filter"
+msg="filter_2"
 rm -f $a.out
 size=$(($good_size+$pro_header+$idx_size+$idx_header-1))
 echo    "$mapgd proview -n $name -H $header | \ "
 echo -n "$mapgd allele -c 1 -b | $mapgd $a -X 0.05 > $a.out								"
 $timeout 5s bash -c "$mapgd proview -n $name -H $header | $mapgd allele -c 1 -b | $mapgd $a -X 0.05 > $a.out			"
-testa
+testc
 
 a="sam2idx"
-msg="sam2idx"
+msg="sam2idx_1"
 rm -f $a.out
 size=$(($idx_size+$idx_header))
 echo -n "cat $header | $mapgd $a > $a.out										"
 $timeout 5s bash -c "cat $header | $mapgd $a > $a.out"
-testa
+testc
 
 a="sam2idx"
-msg="sam2idx"
+msg="sam2idx_2"
 rm -f $a.out
 size=$(($idx_size+$idx_header))
 echo -n "$mapgd $a -H $header > $a.out										"
 $timeout 5s bash -c "$mapgd $a -H $header > $a.out"
-testa
+testc
 
 a="genotype"
-msg="genotype"
+msg="genotype_1"
 rm -f $a.out
 size=$(($idx_size+$idx_header+$good_size+$gcf_header))
 echo "$mapgd proview -H $header -i $mpileup -o temp	 								"
@@ -191,11 +241,11 @@ echo "$mapgd filter -i temp.map -o temp-filtered	 								"
 $mapgd filter -i temp.map -o temp-filtered 
 echo -n "$mapgd $a -p temp.pro -m temp-filtered.map > $a.out								"
 $timeout 5s bash -c "$mapgd $a -p temp.pro -m temp-filtered.map > $a.out"
-testa
+testc
 rm temp*
 
 a="genotype"
-msg="genotype"
+msg="genotype_2"
 rm -f $a.out
 size=$(($idx_size+$idx_header+$good_size+$gcf_header))
 echo "$mapgd proview -H $header -i $mpileup > temp_pro.out 								"
@@ -206,10 +256,11 @@ echo "$mapgd filter -i temp_map.out > temp_map_filtered.out 								"
 $mapgd filter -i temp_map.out > temp_map_filtered.out
 echo -n "$mapgd $a -p temp_pro.out -m temp_map_filtered.out > $a.out							"
 $timeout 5s bash -c "$mapgd $a -p temp_pro.out -m temp_map_filtered.out > $a.out"
-testa
+testc
 rm temp*
 
 echo -n "$mapgd $a -p pro -m map > $a.out (fifopipes)									"
+msg="genotype_3"
 rm -f map
 rm -f pro
 mkfifo map
@@ -219,7 +270,7 @@ $mapgd proview -H $header -i $mpileup | tee pro | $mapgd allele -c 1 | $mapgd fi
 $mapgd genotype -p pro -m map > $a.out
 rm -f map
 rm -f pro
-testa
+testc
 
 a="linkage"
 msg="linkage"
@@ -230,13 +281,13 @@ $mapgd allele -i temp_pro -c 1 > temp_map
 $mapgd filter -i temp_map -p 5 > temp_filtered_map
 $mapgd genotype -p temp_pro -m temp_filtered_map > temp_genotype
 $mapgd linkage -i temp_genotype -M 4 > linkage.out
-testa
+testc
 rm -f temp*
 
 rm -f map
 rm -f pro
 a="relatedness"
-msg="relatedness"
+msg="relatedness_1"
 echo -n "cat genotype.out | $mapgd $a > $a.out 									"
 mkfifo map
 mkfifo pro
@@ -244,13 +295,13 @@ size=$(($pop*($pop-1)/2+3))
 $mapgd proview -H $header -i $mpileup | tee pro | $mapgd allele -c 1 | $mapgd filter > map &
 $mapgd genotype -p pro -m map > genotype
 cat genotype | $mapgd relatedness > $a.out 2> /dev/null
-testa
+testc
 rm -f genotype
 rm -f map
 rm -f pro
 
 a="relatedness"
-msg="relatedness"
+msg="relatedness_2"
 echo -n "cat genotype.out | $mapgd $a -o $a.out 									"
 rm -f map
 rm -f pro
@@ -261,7 +312,7 @@ $mapgd proview -H $header -i $mpileup | tee pro | $mapgd allele -c 1 | $mapgd fi
 $mapgd genotype -p pro -m map -o genotype 
 $mapgd relatedness -i genotype.gcf -o $a.out 2> /dev/null
 mv $a.out.rel $a.out
-testa
+testc
 rm -f $a.idx
 rm -f genotype.gcf
 rm -f genotype.idx
@@ -276,7 +327,7 @@ $mapgd proview -H $header -n $unicode > temp_pro.out
 $mapgd allele -i temp_pro.out -c 1 > temp_allele.out
 $mapgd genotype -p temp_pro.out -m temp_allele.out > temp_genotype.out
 $mapgd relatedness -i temp_genotype.out > $a.out 2> /dev/null
-testa
+testc
 rm -f temp*
 
 a="readphen"
@@ -284,7 +335,7 @@ msg="readphen"
 size=$(($pop+3))
 echo -n "$mapgd $a -i plink.pheno > $a.out 											"
 $mapgd $a -i plink.pheno > $a.out
-testa
+testc
 rm -f temp*
 
 #a="keyinfo"
@@ -349,7 +400,7 @@ rm -f temp*
 
 
 a="write"
-msg="write/read"
+msg="write"
 size=6
 rm -f test.db
 echo "$mapgd sam2idx -H spitze-header.txt | $mapgd write -d test.db 								"
@@ -357,9 +408,9 @@ $mapgd sam2idx -H spitze-header.txt | $mapgd write -d test.db
 echo -n "$mapgd read -d test.db -t REGIONS										"
 $mapgd read -d test.db -t REGIONS > $a.out
 rm -f test.db
-testa
+testc
 
-for com in `$mapgd -a`
-do
-    printf '%-16s\t\t%s\n' $com ${commands[$com]}
-done
+#for com in `$mapgd -a`
+#do
+#    printf '%-16s\t\t%s\n' $com ${commands[$com]}
+#done
